requirements.txt
Django==6.0.4
Pillow==11.3.0
